separate get config and get network
