+++ contract/src/lib.rs
@@ -4,7 +4,6 @@ use btc_types::header::{BlockHeader, ExtendedHeader, Header, LightHeader};
 use btc_types::network::Network;
 use btc_types::u256::U256;
 use btc_types::utils::{target_from_bits, work_from_bits};
-use cfg_if::cfg_if;
 use near_plugins::{
     access_control, pause, AccessControlRole, AccessControllable, Pausable, Upgradable,
 };
@@ -340,42 +339,6 @@ impl BtcLightClient {
                 .mainchain_height_to_header
                 .get(&end_removal_height)
                 .unwrap_or_else(|| env::panic_str(ERR_KEY_NOT_EXIST));
-        }
-    }
-
-    #[cfg(feature = "zcash")]
-    pub fn get_config(&self) -> btc_types::network::ZcashConfig {
-        btc_types::network::get_zcash_config(self.network)
-    }
-
-    #[cfg(feature = "bitcoin")]
-    pub fn get_config(&self) -> btc_types::network::NetworkConfig {
-        btc_types::network::get_bitcoin_config(self.network)
-    }
-
-    #[cfg(feature = "litecoin")]
-    pub fn get_config(&self) -> btc_types::network::NetworkConfig {
-        btc_types::network::get_litecoin_config(self.network)
-    }
-
-    #[cfg(feature = "dogecoin")]
-    pub fn get_config(&self) -> btc_types::network::NetworkConfig {
-        btc_types::network::get_dogecoin_config(self.network)
-    }
-
-    pub fn get_network(&self) -> (String, Network) {
-        cfg_if! {
-            if #[cfg(feature = "bitcoin")] {
-                ("Bitcoin".to_owned(), self.network)
-            } else if #[cfg(feature = "dogecoin")] {
-                ("Dogecoin".to_owned(), self.network)
-            } else if #[cfg(feature = "litecoin")] {
-                ("Litecoin".to_owned(), self.network)
-            } else if #[cfg(feature = "zcash")] {
-                ("Zcash".to_owned(), self.network)
-            } else {
-                compile_error!("No valid network feature enabled.");
-            }
         }
     }
 }
